Preparing for initial release.
Most documentation needs updating according to the recent refactoring.
Changed SimpleRelationManager to a thread safe CopyOnWriteArrayList based implementation (CopyOnWriteRelationManager)

Implemented a new request API. This API allows relation requesters to identify themselves to the resolvers, and acts as a container for all the information concerning the resolution, such as the Entity and the ResolutionContext. A new RelationManager#findRelation(ResolveRequest) method has been added, in favor of this new API.

Slightly improved the code quality of most resolvers, and make PlayerRelationResolver return IGNORED if the player is the local player.
Fixed all resolver implementation tests.